// src/app/models/vehicle.ts
export class VehicleData {
    id?: number;
    car_model?: String;
    car_brand?: String;
    plate_number?: String;
    booking_num?: number;
    car_image?: String;
    available_num?: number;
    broken_down_num?: number;
    model?: String; //NOT AVailable
    marked?: String;
    version?: String;
    category?: string;
    name?: String;
    date_first_circulation?: Date;
    identify_num?: String;
    description?: String;
    capacity?: String;
    price?: String;
    gps_price?: number;
    airbag_price?: number;
    air_conditioning_price?: number;
    child_seat_price?: number;
    cd_player_price?: number;
    baby_seat_price?: number;
    ski_rack?: number;
    rate?: String;
    category_id?: String;
    // 
    dep_city?:String;
    rout_city?:String;
}